Add react-router-dom to project and update App.jsx to use Body component
Adds the "react-router-dom" package to the project and updates App.jsx to use a new Body component. The Body component is imported and used in place of the previous h1 heading. The react and react-dom packages are also updated to version 18.2.0. The tailwind.config.js file is updated to include src files.